+++ app/[locale]/layout.tsx
@@ -5,7 +5,7 @@ import { Toaster } from "sonner";
 import { NextIntlClientProvider } from "next-intl";
 import { notFound } from "next/navigation";
 import { routing } from "@/i18n/routing";
-import { AuthProvider } from "../components/Providers";
+import { AllProviders } from "../components/Providers";
 
 export const metadata: Metadata = {
     title: "Chef",
@@ -25,13 +25,13 @@ export default async function LocaleLayout({ children, params }: { children: Rea
                 <div className="absolute top-0 right-1/4 w-96 h-96 bg-gradient-to-r from-yellow-300/30 to-pink-300/30 rounded-full mix-blend-multiply filter blur-xl opacity-70 animate-blob animation-delay-2000"></div>
                 <div className="absolute bottom-0 left-1/3 w-96 h-96 bg-gradient-to-r from-blue-300/30 to-purple-300/30 rounded-full mix-blend-multiply filter blur-xl opacity-70 animate-blob animation-delay-4000"></div>
             </div>
-            <AuthProvider>
+            <AllProviders>
                 <NextIntlClientProvider messages={messages} locale={locale}>
                     <Header />
                     <Toaster richColors position="top-center" />
                     {children}
                 </NextIntlClientProvider>
-            </AuthProvider>
+            </AllProviders>
         </div>
     );
 }